test and final touch

+++ lib/features/subscription/domain/entities/product_info_entity.dart
@@ -5,7 +5,7 @@ class SubscriptionProductInfo extends Equatable {
   final String? name;
   final int? quantity;
   final num? basePrice;
-  final int? discount;
+  final double? discount;
 
   const SubscriptionProductInfo({
     this.id,
@@ -20,7 +20,7 @@ class SubscriptionProductInfo extends Equatable {
     String? name,
     int? quantity,
     num? basePrice,
-    int? discount,
+    double? discount,
   }) {
     return SubscriptionProductInfo(
       id: id ?? this.id,
